requirements.txt
streamlit>=1.25
pandas>=2.0
scikit-learn>=1.2
xlsxwriter
python-docx
joblib
